+++ pkg/deployment/reconcile/action_encryption_add.go
@@ -27,7 +27,7 @@ import (
 	"context"
 	"encoding/base64"
 
-	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil"
+	"github.com/arangodb/kube-arangodb/pkg/util/globals"
 
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 
@@ -107,7 +107,7 @@ func (a *encryptionKeyAddAction) Start(ctx context.Context) (bool, error) {
 		return true, nil
 	}
 
-	err = k8sutil.RunWithTimeout(ctx, func(ctxChild context.Context) error {
+	err = globals.GetGlobalTimeouts().Kubernetes().RunWithTimeout(ctx, func(ctxChild context.Context) error {
 		_, err := a.actionCtx.SecretsModInterface().Patch(ctxChild, pod.GetEncryptionFolderSecretName(a.actionCtx.GetAPIObject().GetName()), types.JSONPatchType, patch, meta.PatchOptions{})
 		return err
 	})