+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add Recovery during PlanBuild operation
 - Fix Exporter in Deployments without authentication
 - Allow to disable ClusterScalingIntegration and add proper Scheduled label to pods
+- Add additional timeout parameters and kubernetes batch size
 
 ## [1.2.5](https://github.com/arangodb/kube-arangodb/tree/1.2.5) (2021-10-25)
 - Split & Unify Lifecycle management functionality